+++ internal/rest/server_settings.go
@@ -91,7 +91,7 @@ func (s *Server) storeSettings(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	f, verr := settingsform.NewStoreSettingsForm().ParseAndValidate(r)
+	f, verr := settingsform.NewStorePushSettingsForm().ParseAndValidate(r)
 	if verr != nil {
 		response.HandleError(verr, w)
 		return
@@ -111,7 +111,7 @@ func (s *Server) storeSettings(w http.ResponseWriter, r *http.Request) {
 	response.SendEmpty(w, http.StatusOK)
 }
 
-func fillStoreDAOSettingsRequest(in *settingsform.StoreSettingsForm) *inboxapi.PushSettingsDao {
+func fillStoreDAOSettingsRequest(in *settingsform.StorePushSettingsForm) *inboxapi.PushSettingsDao {
 	result := &inboxapi.PushSettingsDao{
 		NewProposalCreated: in.Dao.NewProposalCreated,
 		QuorumReached:      in.Dao.QuorumReached,
@@ -154,6 +154,70 @@ func convertPushDetailsToInternal(in *inboxapi.GetPushDetailsResponse) *settings
 	details.Dao.VoteFinished = in.GetDao().GetVoteFinished()
 	details.Dao.QuorumReached = in.GetDao().GetQuorumReached()
 	details.Dao.NewProposalCreated = in.GetDao().GetNewProposalCreated()
+
+	return details
+}
+
+func (s *Server) storeFeedSettings(w http.ResponseWriter, r *http.Request) {
+	session, ok := appctx.ExtractUserSession(r.Context())
+	if !ok {
+		w.WriteHeader(http.StatusUnauthorized)
+		return
+	}
+
+	f, verr := settingsform.NewStoreFeedSettingsForm().ParseAndValidate(r)
+	if verr != nil {
+		response.HandleError(verr, w)
+		return
+	}
+
+	_, err := s.settings.SetFeedSettings(r.Context(), &inboxapi.SetFeedSettingsRequest{
+		UserId: session.UserID.String(),
+		FeedSettings: &inboxapi.FeedSettings{
+			ArchiveProposalAfterVote: f.ArchiveProposalAfterVote,
+			AutoarchiveAfterDuration: f.AutoarchiveAfterDuration,
+		},
+	})
+	if err != nil {
+		log.Error().Err(err).Msgf("set feed settings for user: %s", session.UserID.String())
+
+		response.SendEmpty(w, http.StatusInternalServerError)
+		return
+	}
+
+	response.SendEmpty(w, http.StatusOK)
+}
+
+func (s *Server) getFeedSettings(w http.ResponseWriter, r *http.Request) {
+	session, ok := appctx.ExtractUserSession(r.Context())
+	if !ok {
+		w.WriteHeader(http.StatusUnauthorized)
+		return
+	}
+
+	details, err := s.settings.GetFeedSettings(r.Context(), &inboxapi.GetFeedSettingsRequest{
+		UserId: session.UserID.String(),
+	})
+	if status.Code(err) == codes.NotFound {
+		response.SendEmpty(w, http.StatusNotFound)
+		return
+	}
+
+	if err != nil {
+		log.Error().Err(err).Msgf("remove push token for user: %s", session.UserID.String())
+
+		response.SendEmpty(w, http.StatusInternalServerError)
+		return
+	}
+
+	response.SendJSON(w, http.StatusOK, convertFeedDetailsToInternal(details))
+}
+
+func convertFeedDetailsToInternal(in *inboxapi.GetFeedSettingsResponse) *settings.FeedSettings {
+	details := &settings.FeedSettings{}
+
+	details.AutoarchiveAfterDuration = in.GetFeedSettings().GetAutoarchiveAfterDuration()
+	details.ArchiveProposalAfterVote = in.GetFeedSettings().GetArchiveProposalAfterVote()
 
 	return details
 }